requirements.txt
bs4
lxml
pillow
pycryptodome
requests
xmltodict